backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
pdf2image==1.16.3
openpyxl==3.1.2
dashscope==1.17.0
requests==2.31.0
bcrypt==4.1.1

